#575 Fix: Removed unused code and moved health-card to backup features
